[WEB-70] refactored the slider to its component
